Added input validation

+++ UserHandler.py
@@ -208,23 +208,25 @@ class UserHandler:
         #takes a string like 'IT 326'
         taken = self.database.get_previous_courses(self.aUser.get_user_name())
         c_Split = course.split()
-        dep = c_Split.pop(0)
-        num = c_Split.pop(0)
-        if self.database.check_for_course(dep,num) and course not in taken:
-            taken.append(course)
-            self.database.add_to_previous_courses(self.aUser.get_user_name(),dep,num)
-            return True
+        if(len(c_Split) == 2):
+            dep = c_Split.pop(0)
+            num = c_Split.pop(0)
+            if self.database.check_for_course(dep,num) and course not in taken:
+                taken.append(course)
+                self.database.add_to_previous_courses(self.aUser.get_user_name(),dep,num)
+                return True
         return False
 
     def remove_previous_course(self, course)->bool: 
         taken = self.database.get_previous_courses(self.aUser.get_user_name())
         c_Split = course.split()
-        dep = c_Split.pop(0)
-        num = c_Split.pop(0)
-        if self.database.check_for_course(dep,num) and course in taken:
-            taken.remove(course)
-            self.database.remove_from_previous_courses(self.aUser.get_user_name(),dep,num)
-            return True
+        if(len(c_Split) == 2):
+            dep = c_Split.pop(0)
+            num = c_Split.pop(0)
+            if self.database.check_for_course(dep,num) and course in taken:
+                taken.remove(course)
+                self.database.remove_from_previous_courses(self.aUser.get_user_name(),dep,num)
+                return True
         return False 
     
     def add_course_to_blacklist(self, course_id, course_dept):
